Add function to ConfFile class
parse_stl_set function is added to ConfFile class to be reused by
multiple users. Also, added overloaded function  extract_options to
extract string value assigned to a key in a conf file
